core/utils: full BigFloat precision
